# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Creator Rewards API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Healthcheck
      responses:
        '200': { description: OK }
  /campaigns:
    get:
      summary: List campaigns
      responses:
        '200': { description: OK }
    post:
      summary: Create campaign (admin)
      parameters:
        - in: header
          name: x-admin-key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                brief: { type: string }
                budgetTotal: { type: integer }
                budgetRemaining: { type: integer }
                ratePerThousand: { type: integer }
                metricType: { type: string, enum: [VIEWS, CLICKS, INTERACTIONS, CONVERSIONS] }
                currency: { type: string }
                brandName: { type: string }
                email: { type: string }
                targetUrl: { type: string }
      responses:
        '201': { description: Created }
        '403': { description: Forbidden }
  /campaigns/{id}:
    get:
      summary: Get campaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
  /campaigns/{id}/performance:
    get:
      summary: Campaign performance
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /campaigns/{id}/applications:
    get:
      summary: List applications for campaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Apply to campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                message: { type: string }
      responses:
        '201': { description: Created }
  /applications:
    get:
      summary: List recent applications (admin)
      parameters:
        - in: header
          name: x-admin-key
          schema: { type: string }
      responses:
        '200': { description: OK }
  /applications/{id}/approve:
    post:
      summary: Approve application (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: header
          name: x-admin-key
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
  /participants/{id}:
    get:
      summary: Get participation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not Found }
  /participants/{id}/submissions:
    post:
      summary: Submit content URL
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platform: { type: string }
                postUrl: { type: string }
                postId: { type: string }
                notes: { type: string }
      responses:
        '201': { description: Created }
  /participants/{id}/metrics:
    get:
      summary: Participant metrics
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /participants/{id}/accrue:
    post:
      summary: Accrue earnings (admin/manual)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: header
          name: x-admin-key
          schema: { type: string }
      responses:
        '200': { description: OK }
  /participants/{id}/lock:
    post:
      summary: Lock eligible accruals (admin/manual)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: header
          name: x-admin-key
          schema: { type: string }
      responses:
        '200': { description: OK }
  /t/{code}:
    get:
      summary: Tracking redirect (records click)
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        '302': { description: Redirect }
  /payments/connect/onboard:
    post:
      summary: Stripe Connect onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200': { description: OK }
  /payouts:
    get:
      summary: List payouts (optional by email)
      parameters:
        - in: query
          name: email
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Create payout from locked accruals (admin)
      parameters:
        - in: header
          name: x-admin-key
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
      responses:
        '200': { description: OK }
components: {}